kaboom

+++ src/main/java/com/simibubi/create/api/behaviour/spouting/BlockSpoutingBehaviour.java
@@ -11,7 +11,7 @@ import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.entity.BlockEntityType;
 import net.minecraft.world.level.block.state.BlockState;
-import net.neoforged.neoforge.fluids.FluidStack;
+import io.github.fabricators_of_create.porting_lib.fluids.FluidStack;
 
 
 /**
@@ -60,5 +60,5 @@ public interface BlockSpoutingBehaviour {
 	 * @param simulate       Whether the spout is testing or actually performing this behaviour
 	 * @return The amount filled into the block, 0 to idle/cancel
 	 */
-	int fillBlock(Level level, BlockPos pos, SpoutBlockEntity spout, FluidStack availableFluid, boolean simulate);
+	long fillBlock(Level level, BlockPos pos, SpoutBlockEntity spout, FluidStack availableFluid, boolean simulate);
 }
